fix(types): resolve all 18 mypy errors
- benchmark.py: dict[str, tuple[float,float]], dict[str, dict[str, Any]],
  TYPE_CHECKING import for QuantumGenesis, explicit float() casts
- system.py: list[dict[str,Any]], dict[str,float] return types on all
  Diamond interface methods; from typing import Any
- qec_threshold.py: float(max/min(...)) to fix no-any-return;
  list[dict[str,Any]] on threshold_scan; from typing import Any
- stim_interface.py: float(min(...)) to fix no-any-return

+++ src/quantum_genesis/stim_interface.py
@@ -68,7 +68,7 @@ class StimInterface:
         """Fowler et al. 2012 analytic approximation."""
         ratio = p / P_THRESHOLD
         exponent = (d + 1) / 2
-        return min(1.0, 0.1 * ratio**exponent)
+        return float(min(1.0, 0.1 * ratio**exponent))
 
     @property
     def backend(self) -> str:
